Refactoring, fixes, improvements

+++ src/main/java/com/loopme/service/FacadeServiceImpl.java
@@ -1,29 +1,30 @@
 package com.loopme.service;
 
-import com.loopme.repository.AppRepository;
-import com.loopme.repository.UserRepository;
-import com.loopme.domain.App;
-import com.loopme.domain.AppType;
-import com.loopme.domain.ContentType;
-import com.loopme.domain.User;
-import com.loopme.domain.UserRole;
-import com.loopme.exception.UserNotFoundException;
-import com.loopme.utils.Utils;
+import java.util.List;
+
+import com.loopme.exception.ApplicationNotFoundException;
+import com.loopme.exception.DuplicateConstraintException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import com.loopme.domain.App;
+import com.loopme.domain.AppType;
+import com.loopme.domain.User;
+import com.loopme.domain.UserRole;
+import com.loopme.exception.UserNotFoundException;
+import com.loopme.repository.AppRepository;
+import com.loopme.repository.UserRepository;
+import com.loopme.utils.Utils;
 
 @Service
 public class FacadeServiceImpl implements FacadeService
 {
-  // TODO : check, if email already registered.
   @Autowired
   private UserRepository userRepository;
 
   @Autowired
-  private AppRepository appRepository;
+  private AppRepository  appRepository;
 
   @Override
   public User createPublisher( String name, String email, String password )
@@ -34,13 +35,13 @@ public class FacadeServiceImpl implements FacadeService
   @Override
   public User editPublisher( Integer id, String name, String email )
   {
-    return updateUser(id, name, email);
+    return updateUser( id, name, email );
   }
 
   @Override
   public void deletePublisher( Integer id )
   {
-    deleteUser(id);
+    deleteUser( id );
   }
 
   @Override
@@ -52,51 +53,52 @@ public class FacadeServiceImpl implements FacadeService
   @Override
   public User editOperator( Integer id, String name, String email )
   {
-    return updateUser(id, name, email);
+    return updateUser( id, name, email );
   }
 
   @Override
   public void deleteOperator( Integer id )
   {
-    deleteUser(id);
+    deleteUser( id );
   }
 
-  // TODO : check input params
   @Override
   public App createApp( String name, String type, List<String> contentTypes )
   {
     String userName = SecurityContextHolder.getContext().getAuthentication().getName();
-    User user = userRepository.getUserByName(userName);
-    App app = new App().setName(name).setUser(user).setType(AppType.valueOf(type)).setContentTypes(Utils.transformToContentTypes(contentTypes));
-    return appRepository.save(app);
+    User user = userRepository.getUserByName( userName );
+    App app = new App().setName( name ).setUser( user ).setType( AppType.valueOf( type ) ).setContentTypes( Utils.transformToContentTypes( contentTypes ) );
+    return appRepository.save( app );
   }
 
-  // TODO : check input params
   @Override
-  public App editApp( Integer id, String name,  String type, List<String> contentTypes )
+  public App editApp( Integer id, String name, String type, List<String> contentTypes )
   {
-    App app = appRepository.getAppById(id);
+    App app = appRepository.getAppById( id );
     if( app == null )
     {
-      // TODO : change exception type
-      throw new UserNotFoundException( "Cannot update user. User with id=" + id + " not found" );
+      throw new ApplicationNotFoundException( "Cannot update application. Application with id=" + id + " not found" );
     }
-    app.setName(name).setType(AppType.valueOf(type)).setContentTypes(Utils.transformToContentTypes(contentTypes));
+    app.setName( name ).setType( AppType.valueOf( type ) ).setContentTypes(Utils.transformToContentTypes(contentTypes));
 
-    return appRepository.save(app);
+    return appRepository.save( app );
   }
 
-  // TODO : check input params
   @Override
   public void deleteApp( Integer id )
   {
-    App app = appRepository.getAppById(id);
-    appRepository.delete(app);
+    App app = appRepository.getAppById( id );
+    if( app == null )
+    {
+      throw new ApplicationNotFoundException( "Cannot delete application. Application with id=" + id + " not found" );
+    }
+    appRepository.delete( app );
   }
 
   @Override
-  public List<User> getPublishers() {
-    return userRepository.getUserByUserRole(UserRole.PUBLISHER);
+  public List<User> getPublishers()
+  {
+    return userRepository.getUserByUserRole( UserRole.PUBLISHER );
   }
 
   @Override
@@ -106,8 +108,9 @@ public class FacadeServiceImpl implements FacadeService
   }
 
   @Override
-  public User getUserByName(String name) {
-    User user = userRepository.getUserByName(name);
+  public User getUserByName( String name )
+  {
+    User user = userRepository.getUserByName( name );
     if( user == null )
     {
       throw new UserNotFoundException( "Cannot update user. User with name : " + name + " not found" );
@@ -116,24 +119,32 @@ public class FacadeServiceImpl implements FacadeService
   }
 
   @Override
-  public List<App> getApps() {
+  public List<App> getApps()
+  {
     String userName = SecurityContextHolder.getContext().getAuthentication().getName();
-    User user = userRepository.getUserByName(userName);
-    if(user.getUserRole() == UserRole.ADOPS){
+    User user = userRepository.getUserByName( userName );
+    if( user.getUserRole() == UserRole.ADOPS )
+    {
       return appRepository.findAll();
     }
-    return appRepository.getAppByUserName(userName);
+    return appRepository.getAppByUserName( userName );
   }
 
-  private void deleteUser(Integer id){
+  private void deleteUser( Integer id )
+  {
     User user = userRepository.getUserById(id);
-    userRepository.delete(user);
+    userRepository.delete( user );
   }
 
   private User createUser( String name, String email, UserRole userRole, String password )
   {
-    Utils.validateEmail( email );
-    User publisherUser = new User().setName( name ).setEmail( email ).setUserRole( userRole ).setPassword(password);
+    Utils.validateEmail(email);
+    User user = userRepository.getUserByName(name);
+    if(user != null)
+    {
+      throw new DuplicateConstraintException("User with name '"+name+"' already exist. Please, try another name.");
+    }
+    User publisherUser = new User().setName( name ).setEmail( email ).setUserRole( userRole ).setPassword( password );
     return userRepository.save( publisherUser );
   }
